rafactor: main mongodb connection class has been created
