+++ services/metrics/app/api/deps.py
@@ -1,6 +1,11 @@
-from app.core.storage import CVMetricManager
+from app.core.training_cv import CVMetricManager
+from app.config import mongodb_config
 
-manager = CVMetricManager()
+manager = CVMetricManager(
+    url=mongodb_config.URL_METRIC,
+    database_name=mongodb_config.DATABASE_METRIC,
+    collection_name=mongodb_config.COLLECTION_TRAINING_CV
+)
 
 async def get_metrics_manager():
     try:
